+++ bika/lims/upgrade/v01_03_000.py
@@ -10,9 +10,27 @@ from bika.lims import logger
 from bika.lims.config import PROJECTNAME as product
 from bika.lims.upgrade import upgradestep
 from bika.lims.upgrade.utils import UpgradeUtils
+from plone.portlets.interfaces import IPortletType
+from zope import component
 
 version = '1.3.0'  # Remember version number in metadata.xml and setup.py
 profile = 'profile-{0}:default'.format(product)
+
+PORTLETS_TO_PURGE = [
+    'accreditation-pt',
+    'login',
+    'news',
+    'events',
+    'Calendar',
+    'portlet_department_filter-pt',
+    'portlet_to_be_sampled-pt',
+    'portlet_to_be_preserved-pt',
+    'portlet_late_analyses-pt',
+    'portlet_pending_orders-pt',
+    'portlet_sample_due-pt',
+    'portlet_to_be_verified-pt',
+    'portlet_verified-pt'
+]
 
 
 @upgradestep(product, version)
@@ -35,6 +53,10 @@ def upgrade(tool):
     # https://github.com/senaite/senaite.core/pull/1058
     remove_qc_reports(portal)
 
+    # Remove old portlets except the navigation portlet
+    # https://github.com/senaite/senaite.core/pull/1060
+    purge_portlets(portal)
+
     logger.info("{0} upgraded to version {1}".format(product, version))
     return True
 
@@ -50,3 +72,26 @@ def remove_qc_reports(portal):
             ti.deleteActions([index])
             break
     logger.info("Removing Reports > Quality Control [DONE]")
+
+
+def purge_portlets(portal):
+    """Remove old portlets. Leave the Navigation portlet only
+    """
+    logger.info("Purging portlets ...")
+
+    def remove_portlets(context_portlet):
+        mapping = portal.restrictedTraverse(context_portlet)
+        for key in mapping.keys():
+            if key not in PORTLETS_TO_PURGE:
+                logger.info("Skipping portlet: '{}'".format(key))
+                continue
+            logger.info("Removing portlet: '{}'".format(key))
+            del mapping[key]
+
+    remove_portlets("++contextportlets++plone.leftcolumn")
+    remove_portlets("++contextportlets++plone.rightcolumn")
+
+    # Reimport the portlets profile
+    setup = portal.portal_setup
+    setup.runImportStepFromProfile(profile, 'portlets')
+    logger.info("Purging portlets [DONE]")